move updateCart functionality to helper function
